requirements.txt
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
Flask==1.1.1
future==0.18.2
idna==2.7
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
python-dotenv==0.14.0
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
urllib3==1.23
webexteamssdk==1.6
Werkzeug==2.0.2